refactor: restructure directory and rename a few files
