+++ src/components/Playbooks/Runs/PlaybookRunsSidePanel.tsx
@@ -4,10 +4,7 @@ import { useAtom } from "jotai";
 import { useEffect, useMemo, useState } from "react";
 import { AiOutlineTeam } from "react-icons/ai";
 import { useNavigate } from "react-router-dom";
-import { ConfigItem } from "../../../api/services/configs";
 import { getPlaybookRuns } from "../../../api/services/playbooks";
-import { Topology } from "../../../context/TopologyPageContext";
-import { HealthCheck } from "../../../types/healthChecks";
 import { relativeDateTime } from "../../../utils/date";
 import PillBadge from "../../Badge/PillBadge";
 import CollapsiblePanel from "../../CollapsiblePanel";
@@ -16,8 +13,7 @@ import { InfiniteTable } from "../../InfiniteTable/InfiniteTable";
 import TextSkeletonLoader from "../../SkeletonLoader/TextSkeletonLoader";
 import { refreshButtonClickedTrigger } from "../../SlidingSideBar";
 import Title from "../../Title/title";
-import { PlaybookSpec } from "../Settings/PlaybookSpecsTable";
-import { User } from "../../../api/services/users";
+import { PlaybookRun } from "./PlaybookRunTypes";
 
 type TopologySidePanelProps = {
   panelType: "topology";
@@ -34,49 +30,13 @@ type Props = {
   onCollapsedStateChange?: (isClosed: boolean) => void;
 };
 
-export type PlaybookRunStatus =
-  | "scheduled"
-  | "running"
-  | "cancelled"
-  | "completed"
-  | "failed"
-  | "pending";
-
-export type PlaybookRunAction = {
-  id: string;
-  name?: string;
-  status: PlaybookRunStatus;
-  playbook_run_id?: string;
-  start_time: string;
-  end_time?: string;
-  result?: {
-    stdout?: string;
-    logs?: string;
-    [key: string]: unknown;
-  };
-  error?: string;
-  playbooks?: PlaybookSpec;
-  playbook_id?: string;
-  created_at?: string;
-  created_by?: User;
-  check_id?: string;
-  config_id?: string;
-  component_id?: string;
-  parameters?: Record<string, unknown>;
-  agent_id?: string;
-  component?: Pick<Topology, "id" | "name" | "icon">;
-  check?: Pick<HealthCheck, "id" | "name" | "icon">;
-  config?: Pick<ConfigItem, "id" | "name" | "type" | "config_class">;
-};
-
-const runsColumns: ColumnDef<PlaybookRunAction, any>[] = [
+const runsColumns: ColumnDef<PlaybookRun, any>[] = [
   {
     header: "Name",
     id: "name",
-    accessorKey: "name",
     size: 60,
     cell: ({ row }) => {
-      const name = row.original.name ?? row.original.playbooks?.name;
+      const name = row.original.playbooks?.name;
       return <span>{name}</span>;
     }
   },
